Wrapped primitives restored better
Previously, if a wrapped primitive, say an Integer, was restored by
asking for an Object, the serializer wouldn't know whether to give
a Integer or a Long (because they both get stored as a primitive long).

In the new version, wrapped primitives are stored as objects like any
other object, so they have a __classname value stored with them.